pipecat-backend/requirements.txt
pipecat-ai[deepgram,anthropic,cartesia,websockets]
fastapi
uvicorn[standard]
python-dotenv
